// src/app/character/models/skills-back.dto.ts
// skills-back.dto.ts
export type SkillsToBackDTO = {
  energyWeapons: number;
  isEnergyWeaponsTagSkill: boolean;

  meleeWeapons: number;
  isMeleeWeaponsTagSkill: boolean;

  smallGuns: number;
  isSmallGunsTagSkill: boolean;

  bigGuns: number;
  isBigGunsTagSkill: boolean;

  athletics: number;
  isAthleticsTagSkill: boolean;

  lockpick: number;
  isLockpickTagSkill: boolean;

  speech: number;
  isSpeechTagSkill: boolean;

  sneak: number;
  isSneakTagSkill: boolean;

  explosives: number;
  isExplosivesTagSkill: boolean;

  unarmed: number;
  isUnarmedTagSkill: boolean;

  medicine: number;
  isMedicineTagSkill: boolean;

  pilot: number;
  isPilotTagSkill: boolean;

  throwing: number;
  isThrowingTagSkill: boolean;

  repair: number;
  isRepairTagSkill: boolean;

  science: number;
  isScienceTagSkill: boolean;

  survival: number;
  isSurvivalTagSkill: boolean;

  barter: number;
  isBarterTagSkill: boolean;
};